requirements.txt
streamlit
pandas
plotly
yfinance
google-genai
feedparser